+++ app/views/documents_view.py
@@ -1,12 +1,10 @@
 import logging
 
 from flet import (
-    AlertDialog,
     ButtonStyle,
     Colors,
     Column,
     Container,
-    Control,
     CrossAxisAlignment,
     Divider,
     ElevatedButton,
@@ -31,8 +29,9 @@ from flet import (
     VerticalDivider,
     alignment,
     border_radius,
-    padding,
 )
+
+from app.views.core import create_modal
 
 logger = logging.getLogger(__name__)
 
@@ -70,20 +69,7 @@ def create_nav_rail_item(page: Page, title: str, id: int, is_authenticated: bool
     )
 
 
-def create_modal(
-    title: Control,
-    content: Control,
-    actions: list[Control],
-    actions_alignment: MainAxisAlignment = MainAxisAlignment.END,
-):
-    return AlertDialog(
-        modal=True,
-        inset_padding=padding.symmetric(vertical=40, horizontal=100),
-        title=title,
-        content=content,
-        actions=actions,
-        actions_alignment=actions_alignment,
-    )
+
 
 
 def create_add_doc_modal(content: TextField, modal_yes_action: callable, modal_no_action: callable):
